backend/requirements.txt
fastapi==0.135.1
uvicorn==0.41.0
modal==1.3.5
supabase==2.28.2
yfinance==1.2.0
python-dotenv==1.2.2
